Lekcja Enum

+++ src/srednieocen/Wynik.java
@@ -1,20 +1,21 @@
 package srednieocen;
 
 public class Wynik {
-    private int ocena;
+    private double ocena;
     private Przedmioty przedmiot;
 
-    public Wynik(int ocena, Przedmioty przedmiot) {
+    public Wynik(double ocena, Przedmioty przedmiot) {
         this.ocena = ocena;
         this.przedmiot = przedmiot;
 
     }
 
-    public int getOcena() {
+    public double getOcena() {
         return ocena;
     }
 
     public Przedmioty getPrzedmiot() {
         return przedmiot;
     }
+
 }
